backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
pynvml
python-dotenv
google-genai
solana
anchorpy
solders